+++ src/neurotechdevkit/scenarios/_scenario_1.py
@@ -216,7 +216,7 @@ class Scenario1_3D(_Scenario1Mixin, Scenario3D):
 
     def get_default_slice_position(self, axis: int) -> float:
         """Return the default slice position for the scenario."""
-        default_positions = np.array([0.064, 0.0, 0.0], dtype=np.float_)
+        default_positions = np.array([0.064, 0.0, 0.0])
         return default_positions[axis]
 
     def render_material_property(
